Fixed #44 (2.5) UnitComposition can now have a priority set to them. UnitCompositionFactory will now return compositions in a prioritized order, with the highest priority being at element [0]. Moved squadId from UnitAgent to BaseAgent and removed the old squadIds. Did some cleanup, some classes in BTHAI are not compiled anymore.

+++ BATS/code/BTHAIModule/Source/RefineryAgent.cpp
@@ -29,7 +29,7 @@ void RefineryAgent::computeActions()
 		}
 	}
 
-	if ((int)assignedWorkers.size() <= 3)
+	if (assignedWorkers.size() <= 3)
 	{
 		if (!unit->isBeingConstructed() && unit->getPlayer()->getID() == Broodwar->self()->getID())
 		{
